+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- LoadingLabel component
+- LoadingProgressBar component
+
 ## [4.1.0] - 2024-12-07
 ### Added
 - SceneManager.QuitGame function